Fix for "Unrecognized switch: -H  (-h will show valid options)" error.
When extra arguments to be passed to script @ARGV like as:

$ speedy -- -gGroup script.pl -H hostname

this "-H" was handled to pass to perl itself arguments, but in this
case, these should be passed as script @ARGV.

If no -g option of speedy is used, speedy_backend is invoked as with
script filename like above, so perl handles -H as script @ARGV.

But with -g option of speedy, script.pl is converted as internal
subroutine call like as:

$ perl -e&{$CGI::SpeedyCGI::_sub}(@ARGV) -H hostname

Then perl trying to treat -H as perl's option, and error occurs.

To avoid this, this patch inserts "--" before rest of arguments.